Introducing text flow #154
This PR introduces the concept of text flow; mulitple calls to 'text' to build a final textBox. While 'flow' is activated, the text position is monitored for user so they don't have to this for themselves. This PR also supports all the ISO named page sizes. Documentation has been updated along with tests.